Issue277: First wave of API restrictions and restructuring
