+++ plugin/src/main/java/org/jenkinsci/plugins/casc/Config.java
@@ -1,7 +1,11 @@
 package org.jenkinsci.plugins.casc;
 
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.Beta;
+
 /**
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
+@Restricted(Beta.class)
 public class Config {
 }
