+++ plugin/src/main/java/org/jenkinsci/plugins/casc/ConfigurationAsCode.java
@@ -1,5 +1,6 @@
 package org.jenkinsci.plugins.casc;
 
+import com.google.common.annotations.VisibleForTesting;
 import hudson.Extension;
 import hudson.init.InitMilestone;
 import hudson.init.Initializer;
@@ -82,6 +83,7 @@ import static org.yaml.snakeyaml.DumperOptions.ScalarStyle.PLAIN;
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
 @Extension
+@Restricted(NoExternalUse.class)
 public class ConfigurationAsCode extends ManagementLink {
 
     public static final String CASC_JENKINS_CONFIG_PROPERTY = "casc.jenkins.config";
@@ -280,7 +282,9 @@ public class ConfigurationAsCode extends ManagementLink {
         }
     }
 
-    static void serializeYamlNode(Node root, Writer writer) throws IOException {
+    @VisibleForTesting
+    @org.kohsuke.accmod.Restricted(NoExternalUse.class)
+    public static void serializeYamlNode(Node root, Writer writer) throws IOException {
         DumperOptions options = new DumperOptions();
         options.setDefaultFlowStyle(BLOCK);
         options.setDefaultScalarStyle(PLAIN);
@@ -293,7 +297,10 @@ public class ConfigurationAsCode extends ManagementLink {
         serializer.close();
     }
 
-    @CheckForNull Node toYaml(CNode config) throws ConfiguratorException {
+    @CheckForNull
+    @VisibleForTesting
+    @org.kohsuke.accmod.Restricted(NoExternalUse.class)
+    public Node toYaml(CNode config) throws ConfiguratorException {
 
         if (config == null) return null;
