+++ plugin/src/main/java/org/jenkinsci/plugins/casc/cli/CheckConfigurationCommand.java
@@ -5,6 +5,8 @@ import hudson.cli.CLICommand;
 import jenkins.model.Jenkins;
 import org.jenkinsci.plugins.casc.ConfigurationAsCode;
 import org.jenkinsci.plugins.casc.yaml.YamlSource;
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.NoExternalUse;
 
 import java.io.InputStream;
 import java.util.Collections;
@@ -13,6 +15,7 @@ import java.util.Collections;
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
 @Extension
+@Restricted(NoExternalUse.class)
 public class CheckConfigurationCommand extends CLICommand {
 
     @Override
