+++ plugin/src/main/java/org/jenkinsci/plugins/casc/core/HudsonPrivateSecurityRealmConfigurator.java
@@ -4,7 +4,9 @@ import hudson.Extension;
 import hudson.security.HudsonPrivateSecurityRealm;
 import org.jenkinsci.plugins.casc.Attribute;
 import org.jenkinsci.plugins.casc.DataBoundConfigurator;
-import org.jenkinsci.plugins.casc.MultivaluedAttribute;
+import org.jenkinsci.plugins.casc.impl.attributes.MultivaluedAttribute;
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.NoExternalUse;
 import org.kohsuke.stapler.DataBoundConstructor;
 
 import java.util.Set;
@@ -14,6 +16,7 @@ import java.util.stream.Collectors;
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
 @Extension
+@Restricted(NoExternalUse.class)
 public class HudsonPrivateSecurityRealmConfigurator extends DataBoundConfigurator {
 
     public HudsonPrivateSecurityRealmConfigurator() {
