+++ plugin/src/main/java/org/jenkinsci/plugins/casc/credentials/SystemCredentialsProviderConfigurator.java
@@ -5,10 +5,11 @@ import com.cloudbees.plugins.credentials.domains.DomainCredentials;
 import hudson.Extension;
 import org.jenkinsci.plugins.casc.Attribute;
 import org.jenkinsci.plugins.casc.BaseConfigurator;
-import org.jenkinsci.plugins.casc.ConfiguratorException;
-import org.jenkinsci.plugins.casc.MultivaluedAttribute;
+import org.jenkinsci.plugins.casc.impl.attributes.MultivaluedAttribute;
 import org.jenkinsci.plugins.casc.model.CNode;
 import org.jenkinsci.plugins.casc.model.Mapping;
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.NoExternalUse;
 
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
@@ -19,6 +20,7 @@ import java.util.Set;
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
 @Extension(optional = true)
+@Restricted(NoExternalUse.class)
 public class SystemCredentialsProviderConfigurator extends BaseConfigurator<SystemCredentialsProvider> {
 
     @Override
