+++ plugin/src/main/java/org/jenkinsci/plugins/casc/integrations/rolestrategy/RoleBasedAuthorizationStrategyConfigurator.java
@@ -1,4 +1,4 @@
-package org.jenkinsci.plugins.casc.integrations.rolebasedauth;
+package org.jenkinsci.plugins.casc.integrations.rolestrategy;
 
 
 import com.michelin.cio.hudson.plugins.rolestrategy.Role;
@@ -9,10 +9,9 @@ import hudson.Extension;
 import org.jenkinsci.plugins.casc.Attribute;
 import org.jenkinsci.plugins.casc.Configurator;
 import org.jenkinsci.plugins.casc.ConfiguratorException;
-import org.jenkinsci.plugins.casc.MultivaluedAttribute;
+import org.jenkinsci.plugins.casc.impl.attributes.MultivaluedAttribute;
 import org.jenkinsci.plugins.casc.model.CNode;
 import org.jenkinsci.plugins.casc.model.Mapping;
-import org.jenkinsci.plugins.casc.model.Sequence;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
