+++ plugin/src/main/java/org/jenkinsci/plugins/casc/plugins/Plugins.java
@@ -1,12 +1,15 @@
 package org.jenkinsci.plugins.casc.plugins;
 
 import hudson.util.VersionNumber;
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.NoExternalUse;
 
 import java.util.HashMap;
 
 /**
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
+@Restricted(NoExternalUse.class)
 public class Plugins extends HashMap<String, VersionNumber> {
 
 }
