+++ plugin/src/main/java/org/jenkinsci/plugins/casc/yaml/StreamReaderWithSource.java
@@ -1,5 +1,7 @@
 package org.jenkinsci.plugins.casc.yaml;
 
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.Beta;
 import org.yaml.snakeyaml.reader.StreamReader;
 
 import java.io.IOException;
@@ -9,6 +11,7 @@ import java.lang.reflect.Field;
  * Hack StreamReader to track the source file/url configuration node have been parsed from
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
+@Restricted(Beta.class)
 public class StreamReaderWithSource extends StreamReader {
 
     public StreamReaderWithSource(YamlSource source) throws IOException {
