+++ plugin/src/main/resources/org/jenkinsci/plugins/casc/DescribableAttribute/schema.jelly
@@ -3,7 +3,7 @@
   "schema" : {
     "oneOf": [
   <j:forEach items="${app.getDescriptorList(it.type)}" var="d" varStatus="st">
-    <j:invokeStatic className="org.jenkinsci.plugins.casc.DescribableAttribute" method="getPreferredSymbol" var="symbol">
+    <j:invokeStatic className="org.jenkinsci.plugins.casc.impl.attributes.DescribableAttribute" method="getPreferredSymbol" var="symbol">
       <j:arg value="${d}" type="hudson.model.Descriptor"/>
       <j:arg value="${d.klass.toJavaClass()}"/>
       <j:arg value="${it.type}"/>
