+++ plugin/src/main/resources/org/jenkinsci/plugins/casc/integrations/jobdsl/SeedJobConfigurator/schema.jelly
