+++ plugin/src/test/resources/org/jenkinsci/plugins/casc/core/Primitives.yml
