+++ plugin/src/test/resources/org/jenkinsci/plugins/casc/core/SetEnvironmentVariable.yml
