+++ plugin/src/test/resources/org/jenkinsci/plugins/casc/credentials/SystemCredentialsTest.yml
