+++ plugin/src/test/resources/org/jenkinsci/plugins/casc/integrations/jobdsl/SeedJobTest.yml
