+++ plugin/src/test/resources/org/jenkinsci/plugins/casc/plugins/ProxyConfigTest.yml
